decompose part 2
