+++ src/components/board/task/dateShowEdit.js
@@ -1,0 +1,15 @@
+export const dateShowMarkup = (date, time) => {
+  return (
+    `<fieldset class="card__date-deadline">
+      <label class="card__input-deadline-wrap">
+        <input
+          class="card__date"
+          type="text"
+          placeholder=""
+          name="date"
+          value="${date} ${time}"
+        />
+      </label>
+    </fieldset>`
+  );
+};
