+++ src/components/board/taskslist.js
@@ -1,4 +1,4 @@
-import createTask from "./task.js";
+import createTask from "./task/task.js";
 
 const createTasksList = (tasks) => {
   let tasksList = ``;
